fix some new lando4 things for demo purposes (#158)
* push failing condition to GHA

* push failing condition to GHA part2

* push failing condition to GHA part3

* push failing condition to GHA part4

* push failing condition to GHA part5

* push failing condition to GHA part6

* push failing condition to GHA part7

+++ utils/generate-build-script.js
@@ -1,14 +1,18 @@
 /* eslint-disable max-len */
 'use strict';
 
-module.exports = (contents, user, mount = '/app') => `
+module.exports = (contents, user, gid, socket, mount = '/app') => `
 #!/bin/sh
 # clean up and setup ssh-auth
-rm -rf /run/ssh-${user}.sock
-sudo socat UNIX-LISTEN:/run/ssh-${user}.sock,fork,user=pirog,group=20,mode=777 UNIX-CONNECT:/run/host-services/ssh-auth.sock &
+if [ -S "${socket}" ]; then
+  rm -rf /run/ssh-${user}.sock
+  sudo socat UNIX-LISTEN:/run/ssh-${user}.sock,fork,user=${user},group=${gid},mode=777 UNIX-CONNECT:${socket} &
+fi
 
 # temp stuff for demo purposes
-git config --global --add safe.directory ${mount}
+if command -v git > /dev/null 2>&1; then
+  git config --global --add safe.directory ${mount}
+fi
 
 ${contents}
 `;
